# Cache MAC address to speed up uuid1/0