--- a/mix.exs
+++ b/mix.exs
@@ -19,7 +19,8 @@ defmodule UUID.Mixfile do
 
   # Application configuration.
   def application do
-    []
+    [extra_applications: [:crypto],
+     mod: {UUID.Application, []}]
   end
 
   # List of dependencies.